CTOF: changed tx,ty,tz coordinates in CTOF::hits to track intersection with CTOF cylinder, changed detector variable in FD and CD trajectory bank to byte

+++ reconstruction/cvt/src/main/java/org/jlab/rec/cvt/banks/RecoBankWriter.java
@@ -478,7 +478,7 @@ public class RecoBankWriter {
             for (StateVec stVec : trks.get(i).get_Trajectory()) {
 
                 bank.setShort("id",       k, (short) trks.get(i).get_Id());
-                bank.setShort("detector", k, (byte) stVec.get_SurfaceDetector());
+                bank.setByte("detector", k, (byte) stVec.get_SurfaceDetector());
                 bank.setByte("sector",    k, (byte) stVec.get_SurfaceSector());
                 bank.setByte("layer",     k, (byte) stVec.get_SurfaceLayer());
                 bank.setFloat("x",        k, (float) (stVec.x()/10.));
@@ -530,7 +530,7 @@ public class RecoBankWriter {
             for (StateVec stVec : trks.get(i).get_Trajectory()) {
 
                 bank.setShort("id",       k, (short) trks.get(i).get_Id());
-                bank.setShort("detector", k, (byte) stVec.get_SurfaceDetector());
+                bank.setByte("detector",  k, (byte) stVec.get_SurfaceDetector());
                 bank.setByte("sector",    k, (byte) stVec.get_SurfaceSector());
                 bank.setByte("layer",     k, (byte) stVec.get_SurfaceLayer());
                 bank.setFloat("x",        k, (float) (stVec.x()/10.));
